+++ PHMS.WebAPI/Application/Application.csproj
@@ -22,4 +22,8 @@
     <ProjectReference Include="..\Domain\Domain.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Commands\Administrator\" />
+  </ItemGroup>
+
 </Project>
