+++ PHMS.WebAPI/Application/Commands/Medic/DeleteMedicCommandById.cs
@@ -1,7 +1,7 @@
 ﻿using Domain.Common;
 using MediatR;
 
-namespace Application.Commands
+namespace Application.Commands.Medic
 {
     public class DeleteMedicCommandById : IRequest<Result<Unit>>
     {
