+++ PHMS.WebAPI/Application/Commands/Medic/UpdateMedicCommandValidator.cs
@@ -1,11 +1,10 @@
-﻿
-using FluentValidation;
+﻿using FluentValidation;
 
-namespace Application.Commands
+namespace Application.Commands.Medic
 {
-    public class UpdatePatientCommandValidator : AbstractValidator<UpdatePatientCommand>
+    public class UpdateMedicCommandValidator : AbstractValidator<UpdateMedicCommand>
     {
-        public UpdatePatientCommandValidator()
+        public UpdateMedicCommandValidator()
         {
             RuleFor(x => x.Id)
                 .NotEmpty().WithMessage("Id is required.")
@@ -50,6 +49,15 @@ namespace Application.Commands
 
             RuleFor(x => x.Address)
                 .NotNull().WithMessage("Address is required.");
+            RuleFor(x => x.Rank)
+                .NotNull().WithMessage("Rank is required.")
+                .MaximumLength(30).WithMessage("Rank must be at most 30 characters.");
+            RuleFor(x => x.Specialization)
+                .NotNull().WithMessage("Specialization is required.")
+                .MaximumLength(30).WithMessage("Specialization must be at most 30 characters.");
+            RuleFor(x => x.Hospital)
+                .NotNull().WithMessage("Hospital is required.")
+                .MaximumLength(30).WithMessage("Hospital must be at most 30 characters.");
 
         }
         private bool BeAValidGuid(Guid guid)
