+++ PHMS.WebAPI/Application/Commands/Patient/UpdatePatientCommand.cs
@@ -1,10 +1,11 @@
 ﻿using Domain.Common;
 using MediatR;
 
-namespace Application.Commands
+namespace Application.Commands.Patient
 {
-    public class CreatePatientCommand : IRequest<Result<Guid>>
+    public class UpdatePatientCommand : IRequest<Result<Unit>>
     {
+        public Guid Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public DateTime BirthDate { get; set; }
