+++ PHMS.WebAPI/Application/Utils/MappingProfile.cs
@@ -1,4 +1,5 @@
-using Application.Commands;
+using Application.Commands.Medic;
+using Application.Commands.Patient;
 using Application.DTOs;
 using AutoMapper;
 using Domain.Entities;
