+++ PHMS.WebAPI/PHMS/Controllers/MedicController.cs
@@ -1,4 +1,5 @@
 ﻿using Application.Commands;
+using Application.Commands.Medic;
 using Application.DTOs;
 using Application.Queries;
 using MediatR;
